--- src/main/java/com/ureca/uhyu/domain/admin/dto/response/UserBrandPair.java
@@ -1,8 +1,0 @@
-package com.ureca.uhyu.domain.admin.dto.response;
-
-public record UserBrandPair(
-        Long userId,
-        Long brandId
-) {
-
-}
